+++ dflib-echarts/src/main/java/org/dflib/echarts/LabelPosition.java
@@ -1,0 +1,10 @@
+package org.dflib.echarts;
+
+/**
+ * @since 1.0.0-M22
+ */
+public enum LabelPosition {
+    top, left, right, bottom,
+    inside, insideLeft, insideRight, insideTop, insideBottom,
+    insideTopLeft, insideBottomLeft, insideTopRight, insideBottomRight
+}