+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #291 Remove APIs deprecated since M20
 * #292 Invalid ECharts JS for strings containing single quotes
 * #293 ECharts: SeriesOps.showSymbol
+* #294 ECharts: SeriesOpts.label
 
 ## Release 1.0.0-M21